Dynamically add top-level submodules
Any module added to the top-level of the abs package, except those
starting with underscores, will automatically be added to the abs
package namespace, so the following will work, without needing to
explicitly import the ascceg module in __init__:

    import abs
    df = abs.ascceg.df()

Additionally, these top-level modules are added to __all__, which
explicitly marks them as public submodules of the package, and makes
them available for `from abs import *`.

Additionally, these top-level modules are automatically added as CLI
subcommands. So you can run the following, without needing to explicitly
add the submodule as a command:

    abs ascceg -f csv